requirements.txt
#Python == 3.9.21
Flask == 3.1.0
Werkzeug == 3.1.3
python-dotenv==1.0.0
requests==2.32.3
urllib3==2.2.3